requirements.txt
flask
flask-login
psycopg2-binary
celery
redis
flask-wtf
flask-sqlalchemy
pip-tools
email_validator
pytest
pytest-cov
flask_mail==0.9.0
connexion[swagger-ui]